feat: cleanup & drop finality http
