--- packages/backend/src/modules/finality/api/FinalityRouter.ts
@@ -1,15 +1,0 @@
-import Router from '@koa/router'
-
-import { FinalityController } from './FinalityController'
-
-export function createFinalityRouter(finalityController: FinalityController) {
-  const router = new Router()
-
-  router.get('/api/finality', async (ctx) => {
-    const result = await finalityController.getFinality()
-
-    ctx.body = result
-  })
-
-  return router
-}
