+++ packages/frontend2/src/app/(new)/(other)/scaling/finality/page.tsx
@@ -2,6 +2,7 @@ import { getDefaultMetadata } from '~/utils/get-default-metadata'
 
 import { layer2s } from '@l2beat/config'
 import { SimplePageHeader } from '~/app/_components/simple-page-header'
+import { getImplementationChangeReport } from '~/server/features/implementation-change-report/get-implementation-change-report'
 import { getFinality } from '~/server/features/scaling/finality/get-finality'
 import { getFinalityConfigurations } from '~/server/features/scaling/finality/get-finality-configurations'
 import { getScalingFinalityEntries } from '~/server/features/scaling/finality/get-scaling-finality-entries'
@@ -20,12 +21,18 @@ export const metadata = getDefaultMetadata({
 export default async function Page() {
   const configurations = getFinalityConfigurations()
 
-  const [finality, tvl] = await Promise.all([
+  const [finality, tvlForOrdering, icReport] = await Promise.all([
     getFinality(configurations),
     getLatestTvlUsd({ type: 'layer2' }),
+    getImplementationChangeReport(),
   ])
 
-  const projects = await getScalingFinalityEntries(layer2s, tvl, finality)
+  const projects = await getScalingFinalityEntries(
+    layer2s,
+    tvlForOrdering,
+    finality,
+    icReport,
+  )
 
   return (
     <>
