+++ packages/frontend2/src/server/features/scaling/finality/calculate-intervals.ts
@@ -1,4 +1,4 @@
-import { type LivenessRecordWithConfig } from './LivenessWithConfigService'
+import { type LivenessRecordWithConfig } from './get-liveness-by-type-since'
 
 export type Interval = {
   record: LivenessRecordWithConfig
