+++ packages/frontend2/src/server/features/scaling/finality/filter-intervals-by-range.test.ts
@@ -1,11 +1,12 @@
 import { UnixTime } from '@l2beat/shared-pure'
 import { expect, mockObject } from 'earl'
-import { type LivenessRecordWithConfig } from './LivenessWithConfigService'
-import { type Interval } from './calculateIntervals'
-import { calculateStats } from './calculateStats'
+import { type Interval } from './calculate-intervals'
+import { filterIntervalsByRange } from './filter-intervals-by-range'
+import { type LivenessRecordWithConfig } from './get-liveness-by-type-since'
 
-describe(calculateStats.name, () => {
-  it('should calculate stats', async () => {
+
+describe(filterIntervalsByRange.name, () => {
+  it('should filter intervals by range', async () => {
     const NOW = UnixTime.now()
     const MOCK_INTERVALS: Interval[] = [
       {
@@ -31,12 +32,8 @@ describe(calculateStats.name, () => {
       },
     ]
 
-    const result = calculateStats(MOCK_INTERVALS)
+    const result = filterIntervalsByRange(MOCK_INTERVALS, NOW, '30D')
 
-    expect(result).toEqual({
-      averageInSeconds: 20,
-      minimumInSeconds: 10,
-      maximumInSeconds: 30,
-    })
+    expect(result).toEqual([...MOCK_INTERVALS].splice(0, 2))
   })
 })
